backend/generated_projects/app-1758468489648/backend/requirements.txt
# ASGI Server
uvicorn[standard]

# Web Framework
fastapi

# For settings management from environment variables
python-dotenv